requirements.txt
alembic==1.7.5
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.6
pytest==6.2.5
pytest-flask==1.2.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
